applied user authentication to todo requests

+++ routers/todos.py
@@ -1,12 +1,18 @@
 from typing import Annotated
+
 from fastapi import Depends, HTTPException, Path, APIRouter
 from pydantic import BaseModel, Field
-from starlette import status
-from TodoApp.database import SessionLocal
 from sqlalchemy.orm import Session
-from TodoApp.models import Todo
+from starlette import status
 
-router = APIRouter()
+from TodoApp.database import SessionLocal
+from TodoApp.models import Todo
+from routers.auth import get_current_user
+
+router = APIRouter(
+    prefix="/todos",
+    tags=["todos"],
+)
 
 
 def get_db():
@@ -18,6 +24,7 @@ def get_db():
 
 
 db_dependency = Annotated[Session, Depends(get_db)]
+user_dependency = Annotated[dict, Depends(get_current_user)]
 
 
 class TodoRequest(BaseModel):
@@ -27,32 +34,47 @@ class TodoRequest(BaseModel):
     complete: bool
 
 
-@router.get("/todos", status_code=status.HTTP_200_OK)
-async def find_all_todos(db: db_dependency):
-    return db.query(Todo).all()
+@router.get("/", status_code=status.HTTP_200_OK)
+async def find_all_todos(db: db_dependency, current_user: user_dependency):
+    if current_user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    return db.query(Todo).filter(Todo.owner_id == current_user.get("id")).all()
 
 
-@router.get("/todos/{todo_id}", status_code=status.HTTP_200_OK)
-async def find_by_id(db: db_dependency, todo_id: Annotated[int, Path(gt=0)]):
-    found_todo = db.query(Todo).filter(Todo.id == todo_id).first()
+@router.get("/{todo_id}", status_code=status.HTTP_200_OK)
+async def find_by_id(db: db_dependency, current_user: user_dependency, todo_id: Annotated[int, Path(gt=0)]):
+    if current_user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    found_todo = db.query(Todo).filter(Todo.id == todo_id and Todo.owner_id == current_user.get("id")).first()
+
     if found_todo is None:
         raise HTTPException(status_code=404, detail="Todo not found")
+
     return found_todo
 
 
-@router.post("/todos", status_code=status.HTTP_201_CREATED)
-async def create_todo(db: db_dependency, body_req: TodoRequest):
+@router.post("/", status_code=status.HTTP_201_CREATED)
+async def create_todo(db: db_dependency, current_user: user_dependency, body_req: TodoRequest):
+    if current_user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
     try:
-        new_todo = Todo(**body_req.model_dump())
+        new_todo = Todo(**body_req.model_dump(), owner_id=current_user.get("id"))
         db.add(new_todo)
         db.commit()
     except Exception as error:
         raise HTTPException(status_code=400, detail=str(error))
 
 
-@router.put("/todos/{todo_id}", status_code=status.HTTP_201_CREATED)
-async def update_todo(db: db_dependency, todo_id: Annotated[int, Path(gt=0)], body_req: TodoRequest):
-    found_todo = db.query(Todo).filter(Todo.id == todo_id).first()
+@router.put("/{todo_id}", status_code=status.HTTP_201_CREATED)
+async def update_todo(db: db_dependency, current_user: user_dependency, todo_id: Annotated[int, Path(gt=0)],
+                      body_req: TodoRequest):
+    if current_user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    found_todo = db.query(Todo).filter(Todo.id == todo_id and Todo.owner_id == current_user.get("id")).first()
 
     if found_todo is None:
         raise HTTPException(status_code=404, detail="TODO not found")
@@ -66,13 +88,16 @@ async def update_todo(db: db_dependency, todo_id: Annotated[int, Path(gt=0)], bo
     db.commit()
 
 
-@router.delete("/todos/{todo_id}", status_code=status.HTTP_201_CREATED)
-async def delete_todo(db: db_dependency, todo_id: Annotated[int, Path(gt=0)], body_req: TodoRequest):
-    found_todo = db.query(Todo).filter(Todo.id == todo_id).first()
+@router.delete("/{todo_id}", status_code=status.HTTP_201_CREATED)
+async def delete_todo(db: db_dependency, current_user: user_dependency, todo_id: Annotated[int, Path(gt=0)]):
+    if current_user is None:
+        raise HTTPException(status_code=401, detail="Authentication failed")
+
+    found_todo = db.query(Todo).filter(Todo.id == todo_id and Todo.owner_id == current_user.get("id")).first()
 
     if found_todo is None:
         raise HTTPException(status_code=404, detail="TODO not found")
 
-    db.query(Todo).filter(Todo.id == todo_id).delete()
+    db.query(Todo).filter(Todo.id == todo_id and current_user.get("id") == Todo.owner_id).delete()
 
     db.commit()
